refactor(web): decompose the request Logs page
RequestLogs was one 907-line component. It keeps the state, the
paginate query, the cursor fetch, wheel paging and the page composition
(394 lines, under the max-lines-per-function cap, so its eslint-disable
goes); the rest moves under pages/Logs/ next to utils.ts:

- useRequestLogLiveUpdates.ts: the paginate-mode invalidation and scroll
  mode's SSE merge-by-id handler, 60s poll and visibility refresh; also
  reports document visibility for the paginate refetch interval
- useStaleClock.ts: parseGoDuration, the stale threshold from the
  setting, the 1s/60s ticking clock, isStale/isInProgress
- RequestLogRow.tsx: one request row of the paginated table
- RequestLogFilters.tsx: the controls card (sub-mode switch, view
  toggle, filters, date picker)
- requestLogColumns.ts: LogSortField and the twelve sortable columns as
  data, built from t so every label stays a literal t() call; the
  header row maps over them instead of twelve hand-written headers

No behaviour change: same markup, query keys and event handling; the 11
Logs suites pass unchanged. The file leaves the size allowlist
(971 -> 394).

Verified on the rebuilt dev stack: scroll mode lists the virtual rows,
paginate mode renders the 12 headers and 20 rows, a row opens the
detail modal, a no-match filter shows the empty row, the App Logs
sub-mode switches, no console errors.
